+++ src/components/index.js
@@ -4,6 +4,7 @@ import Button from 'components/Button';
 import ButtonGroup from 'components/ButtonGroup';
 import Checkbox from 'components/Checkbox';
 import Chip from 'components/Chip';
+import CornerPosition from 'components/CornerPosition';
 import DateInput from 'components/DateInput';
 import Dialog from 'components/Dialog';
 import DropdownButton from 'components/DropdownButton';
@@ -23,7 +24,6 @@ import StackedButton from 'components/StackedButton';
 import TextInput from 'components/TextInput';
 import Toaster from 'components/Toaster';
 import Tooltip from 'components/Tooltip';
-import Updatable from 'components/Updatable';
 import ZeroState from 'components/ZeroState';
 
 export {
@@ -33,6 +33,7 @@ export {
   ButtonGroup,
   Checkbox,
   Chip,
+  CornerPosition,
   DateInput,
   Dialog,
   DropdownButton,
@@ -52,6 +53,5 @@ export {
   TextInput,
   Toaster,
   Tooltip,
-  Updatable,
   ZeroState,
 };